+++ hodu-cli/src/output.rs
@@ -185,10 +185,10 @@ pub fn format_duration(secs: f64) -> String {
 }
 
 /// Format byte size in human readable form
-pub fn format_size(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = KB * 1024;
-    const GB: u64 = MB * 1024;
+pub fn format_size(bytes: usize) -> String {
+    const KB: usize = 1024;
+    const MB: usize = KB * 1024;
+    const GB: usize = MB * 1024;
 
     if bytes >= GB {
         format!("{:.2} GB", bytes as f64 / GB as f64)
